README.md
no code for ESP8266 with DHT22 sensor for a web thermostat application running on a NODE MCU board.

A low cost NODE MCU ESP8266 board with a lower cost additional DHT22 sensor are used as a thermostat to control the central heating of a house. A relay is also needed to use the low voltage from the board to control the high voltage which drivs the heating system.

The server runs solely on the board and is accessible by any browser. User can determin if heating is on or off and the desired temperature. Interface text is in Greek, though the comments in the code are in english.